+++ Runtime/physics/tolerance.cs
@@ -1,24 +1,24 @@
 namespace Unity.UIWidgets.physics {
     public class Tolerance {
         public Tolerance(
-            double distance = _epsilonDefault,
-            double time = _epsilonDefault,
-            double velocity = _epsilonDefault
+            float distance = _epsilonDefault,
+            float time = _epsilonDefault,
+            float velocity = _epsilonDefault
         ) {
             this.distance = distance;
             this.time = time;
             this.velocity = velocity;
         }
 
-        const double _epsilonDefault = 1e-3;
+        const float _epsilonDefault = 1e-3f;
 
         public static readonly Tolerance defaultTolerance = new Tolerance();
 
-        public readonly double distance;
+        public readonly float distance;
 
-        public readonly double time;
+        public readonly float time;
 
-        public readonly double velocity;
+        public readonly float velocity;
 
         public override string ToString() {
             return $"Tolerance(distance: ±{this.distance}, time: ±{this.time}, velocity: ±{this.velocity})";
