+++ Runtime/physics/utils.cs
@@ -2,7 +2,7 @@ using Unity.UIWidgets.foundation;
 
 namespace Unity.UIWidgets.physics {
     public class PhysicsUtils {
-        public static bool nearEqual(double? a, double? b, double epsilon) {
+        public static bool nearEqual(float? a, float? b, float epsilon) {
             D.assert(epsilon >= 0.0);
             if (a == null || b == null) {
                 return a == b;
@@ -11,8 +11,8 @@ namespace Unity.UIWidgets.physics {
             return (a > (b - epsilon)) && (a < (b + epsilon)) || a == b;
         }
 
-        public static bool nearZero(double? a, double epsilon) {
-            return nearEqual(a, 0.0, epsilon);
+        public static bool nearZero(float? a, float epsilon) {
+            return nearEqual(a, 0.0f, epsilon);
         }
     }
 }
